Refactored the Synchronizbale interface a bit to make the param a long long and to add a new wait method that takes a long long and an int value for nanoseconds.  Added a start on a PriorityQueue implementation.  Added a new Timer class that mimics the Java Timer class and added the TimerTask interface that goes along with it.  Performed some other API documentation cleanup.
git-svn-id: https://svn.apache.org/repos/asf/activemq/activemq-cpp/trunk@796317 13f79535-47bb-0310-9956-ffa450edef68

+++ activemq-cpp/src/main/decaf/internal/io/StandardInputStream.h
@@ -157,7 +157,7 @@ namespace io {
             mutex.wait();
         }
 
-        virtual void wait( unsigned long millisecs ) throw( lang::Exception ){
+        virtual void wait( long long millisecs ) throw( lang::Exception ){
             mutex.wait( millisecs );
         }
